Remove panic unwinding and optimize build size
For the purpose of code cleanup and optimiziing file sizes, we remove
the panic unwinding logic.

There doesn't remain any spot in the code that will propagate a rust
panic. Any panics that happen in Tokio are already propagated as errors.
Any sort of invalid state that happens in the hooks are either logically
unreachable code (validated in test and in runtime), or are in unsafe
code.

In the case that an error happens in unsafe code, it should for the most
part be gracefully handled in logging. In serious memory access
violations, the program will crash without invoking a Rust panic anyway.